chore: refactoring ViewMap toolbars
