+++ nerdlets/geo-ops-nerdlet/ViewMap/Toolbars/RightToolbar.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Button } from 'nr1';
 
-import { ToolbarItem } from '../../../shared/components';
+import { ToolbarItem } from '../../components';
 
 const RightToolbar = ({ navigation }) => {
   return (
